Update Preact types accordingly.
